# perf(functions): Optimize Min/Max construction with many arguments (fixes #16249)