--- a/.mailmap
+++ b/.mailmap
@@ -1047,6 +1047,7 @@ Miro Hrončok <miro@hroncok.cz>
 Mivik <mivikq@gmail.com> Mivik <54128043+Mivik@users.noreply.github.com>
 Mivik <mivikq@gmail.com> Mivik <mivikq@gmail.com>
 Mohak Malviya <mohakmalviya2000@gmail.com> Senku <mohakmalviya2000@gmail.com>
+Mohamed Laarej <mohamed.laarej.eng@gmail.com>
 Mohamed Rezk <mohrizq895@gmail.com>
 Mohammad Sadeq Dousti <msdousti@gmail.com>
 Mohammed Bilal <r.mohammedbilal@gmail.com>