@mohamed-laarej 2025-03-14
The failing test test_issue_21890 is unrelated to my changes and was already failing in the main branch. The error is about comparing complex numbers (TypeError: Invalid comparison of non-real 2**(2/3)*sqrt(3)*I/(2*x)), which is not affected by my Min/Max performance optimizations.
@oscarbenjamin Would you like me to proceed with this PR focused on the Min/Max performance optimization, or would you prefer I look into fixing the test_issue_21890 failure separately? I'm happy to help with either approach.
For reference, my benchmark results show significant performance improvements:
 - Min with 50 symbols (evaluate=True): 3.24s → 0.99s (3.3x faster)
 - Max with 50 symbols (evaluate=True): 3.23s → 0.94s (3.4x faster)
 
evaluate=False performance remains fast as expected