@oscarbenjamin 2025-04-14
I think that the proper fix here is really for `Max` and `Min` to be much more conservative about attempting to simplify anything.