@oscarbenjamin 2025-04-14
> The failing test test_issue_21890 is unrelated to my changes and was already failing in the main branch.

That failure is definitely caused by the changes here. The traceback clearly shows the code here as the cause.